+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
  ## [Unreleased]
  [Added]
  - List of services is displayed as JSON if parameter 'output=json' is set in URL
- 
+
  [Fixed]
  - Fixed the problem where LDAP calls RPC method in PerunIdentity filter
+ - Fixed assignation of one Perun attribute to multiple SP attributes
  
  ## [v2.1.0]
  [Added]